+++ modules/data_reading/data_reading.py
@@ -16,90 +16,63 @@ from configs import args, logging, PATH
 log = logging.getLogger("spacy")
 log.setLevel(logging.ERROR)
 
-nlp             = spacy.load("en_core_web_sm", disable=['ner', 'parser', 'tagger'])
-nlp.max_length  = 2500000
-
-MAX_LEN_PARA    = 50
-MAX_LEVEL       = 3
-MAX_NEIGHBORS   = 3
+nlp             = spacy.load("en_core_web_sm")
+nlp.add_pipe('sentencizer')
 
 
-def clean_text(para:str) -> str:
-    para = re.sub(r'(\n|\t)', '', para)
-    para = re.sub(r'\s{2,}', ' ', para)
+class DataReading():
+    def __init__(self, split) -> None:
+        self.processed_contx    = {}
+        self.split              = split
 
-    para = para.strip()
+    ############################################################
+    # Methods to process context, question and answers
+    ############################################################
+    def clean_end(self, text:str):
+        text = text.split(' ')
+        return ' '.join(text[:-2])
 
-    return para.lower()
+    def clean_context_movie(self, context:str) -> str:
+        context = context.lower().strip()
 
-def clean_end(text:str):
-    text = text.split(' ')
-    return ' '.join(text[:-2])
+        context = re.sub(r'( {2,}|\t)', ' ', context)
+        context = re.sub(r' \n ', '\n', context)
+        # context = re.sub(r'\n ', ' ', context)
+        context = re.sub(r"(?<=\w|,)\n(?=\w| )", ' ', context)
+        context = re.sub(r'\n{2,}', '\n', context)
 
-def get_score(v1, v2):
-    if np.linalg.norm(v1) * np.linalg.norm(v2) == 0:
-        return -1
+        return context
 
-    return 1 - spatial.distance.cosine(v1, v2)
+    def clean_context_gutenberg(self, context:str) -> str:
+        context = context.lower().strip()
 
-def find_golden(query, paragraphs, level:int, neighbors:set):
-    # print(f"{level} - {query}")
+        context = re.sub(r"(?<=\w|,|;|-)\n(?=\w| )", ' ', context)
+        context = re.sub(r'( {2,}|\t)', ' ', context)
+        # context = re.sub(r' \n ', '\n', context)
+        # context = re.sub(r'\n ', ' ', context)
+        context = re.sub(r'\n{2,}', '\n', context)
 
+    def process_context_movie(self, context, start, end) -> list:
+        """Process context and split into paragrapgs
 
-    ## if level not exceed MAX_LEVEL, keep calling function recursively
-    if level == MAX_LEVEL:
-        return
+        Args:
+            context (str): context (book or movie script)
+            start (str): start word of context
+            end (str): end word of context
 
-    ## Calculate score of query for each para
-    scores = [
-        (ith, get_score(paragraphs[query], paragraphs[ith]))
-        for ith in range(0, len(paragraphs) - 3)
-    ]
+        Returns:
+            list: list of paras
+        """
 
-    ## Sort paras w.r.t relevant score to query
-    scores.sort(reverse=True, key= lambda x: x[1])
-
-    ## while not reaching MAX_NEIGHBORS, keep adding to set if not appearing beforehand in set
-    found_  = []
-    for k in scores:
-        if k[0] not in neighbors:
-            neighbors.add(k[0])
-            found_.append(k[0])
-
-        # NOTE: This line is eddited for analyzing purpose
-        if len(found_) == args.n_paras:
-            break
-    
-    # NOTE: This block is temporarily commented due to analyzing purpose
-    # for f in found_:
-    #     find_golden(f, paragraphs, level+1, neighbors)
-
-def f_trigger(documents, queue):
-    """This function is used to run in parallel to read and initially preprocess
-    raw documents.
-
-    Args:
-        document (dict): a document of dataset
-
-    Returns:
-        dict: a dict containing new fields with value
-    """
-    for document in documents:
-        context = document['document']['text']
-
-        ## Preprocess
         ## Extract text from HTML
         soup    = BeautifulSoup(context, 'html.parser')
         if soup.pre is not None:
             context = ''.join(list(soup.pre.findAll(text=True)))
 
-        ## Clean and lowercase
-        context = clean_text(context)
-
-
-        start_  = document['document']['start'].lower()
-        end_    = document['document']['end'].lower()
-        end_    = clean_end(end_)
+        ## Use field 'start' and 'end' provided
+        start_  = start.lower()
+        end_    = end.lower()
+        end_    = self.clean_end(end_)
 
         start_  = context.find(start_)
         end_    = context.rfind(end_)
@@ -110,81 +83,269 @@ def f_trigger(documents, queue):
 
         context = context[start_:end_]
 
+        ## Clean context and split into paras
+        sentences   = self.clean_context_movie(context).split('\n')
 
-        ## Tokenize
-        context = [tok.text for tok in nlp(context)]
-
-
-        ## Split into paragraphs
-        paras               = []
-        paras_preprocessed  = []
-        for ith in range(0, len(context), MAX_LEN_PARA):
-            para    = ' '.join(context[ith:ith+MAX_LEN_PARA])
-
-            paras.append(para)
-
+        paras       = []
+        n_tok       = 0 
+        para        = ""
+        for sent in sentences:
             ## Tokenize, remove stopword and lemmatization
-            tok_context = []
-            for tok in nlp(para):
-                if not (tok.is_stop or tok.is_punct):
-                    tok_context.append(tok.lemma_)
-            paras_preprocessed.append(' '.join(tok_context))
+            sent_   = [tok.lemma_ for tok in nlp(sent)
+                       if not (tok.is_stop or tok.is_punct)]
+
+            ## Concat sentece if not exceed paragraph max len
+            if n_tok + len(sent_) < 50:
+                n_tok   += len(sent_)
+                para    = para + ' ' + ' '.join(sent_)
+            else:
+                paras.append(para)
+                n_tok   = 0
+                para    = ""
 
 
-        tfidfvectorizer = TfidfVectorizer(analyzer='word', stop_words='english', ngram_range=(1, 3),
-                                      max_features=500000)
-
-        paras_preprocessed.append(document['question']['text'].lower())
-        paras_preprocessed.append(document['answers'][0]['text'].lower())
-        paras_preprocessed.append(document['answers'][1]['text'].lower())
+        return paras
 
 
-        wm = tfidfvectorizer.fit_transform(paras_preprocessed).toarray()
+
+
+        return context
+
+    def process_context_gutenberg(self, context, start, end):
+        soup    = BeautifulSoup(context, 'html.parser')
+        if soup.pre is not None:
+            context = ''.join(list(soup.pre.findAll(text=True)))
+
+        # Use field 'start' and 'end' provided
+        start_  = start.lower()
+        end_    = end.lower()
+        end_    = self.clean_end(end_)
+
+        start_  = context.find(start_)
+        end_    = context.rfind(end_)
+        if start_ == -1:
+            start_ = 0
+        if end_ == -1:
+            end_ = len(context)
+
+        context = context[start_:end_]
+
+        ## Clean context and split into paras
+        sentences   = self.clean_context_gutenberg(context).split('\n')
+
+        paras       = np.array([])
+        n_tok       = 0
+        para        = ""
+        for sent in sentences:
+            ## Tokenize, remove stopword
+            sent_   = [tok.text for tok in nlp(sent)
+                    if not (tok.is_stop or tok.is_punct)]
+
+            if len(sent_) > 50:
+                # Long paragraph: split into sentences and
+                # apply concatenating strategy
+
+                # Finish concateraning para
+                if para != "":
+                    paras = np.append(paras, para)
+                    n_tok   = 0
+                    para    = ""
+
+                # Split into sentences and
+                # apply concatenating strategy
+                for sub_sent in nlp(sent).sents:
+                    sent_   = [tok.text for tok in sub_sent
+                            if not (tok.is_stop or tok.is_punct)]
+
+                    if n_tok + len(sent_) < 50:
+                        n_tok   += len(sent_)
+                        if para != "":
+                            para    = para + ' ' + ' '.join(sent_)
+                        else:
+                            para    = ' '.join(sent_)
+                    else:
+                        paras   = np.append(paras, para)
+                        n_tok   = len(sent_)
+                        para    = ' '.join(sent_)
+
+                if para != "":
+                    paras = np.append(paras, para)
+                    n_tok   = 0
+                    para    = ""
+
+            else:
+                if n_tok + len(sent_) < 50:
+                    n_tok   += len(sent_)
+                    if para != "":
+                        para    = para + ' ' + ' '.join(sent_)
+                    else:
+                        para    = ' '.join(sent_)
+                else:
+                    paras = np.append(paras, para)
+                    n_tok   = 0
+                    para    = ""
+
+        if para != "":
+            paras   = np.append(paras, para)
+            n_tok   = len(sent_)
+            para    = ' '.join(sent_)
+
+        paras = paras.tolist()
+
+        return paras
+
+
+    def process_contx(self, keys, queue):
+        for key in keys:
+            context, kind, start, end = self.processed_contx[key]
+
+            if kind == "movie":
+                paras   = self.process_context_movie(context, start, end)
+            else:
+                paras   = self.process_context_gutenberg(context, start, end)
+
+            self.processed_contx[key]   = paras
+
+            queue.put(1)
+
+
+    def process_ques_ans(self, text):
+        """Process question/answers
+
+        Args:
+            text (str): question or answers
+
+        Returns:
+            str: processed result
+        """
+        tok = [tok.lemma_ for tok in nlp(text)
+               if not (tok.is_stop or tok.is_punct)]
+
+        return ' '.join(tok)
+
+    ############################################################
+    # Methods to find golden passages given question or answers
+    ############################################################
+    def get_score(self, v1, v2):
+        if np.linalg.norm(v1) * np.linalg.norm(v2) == 0:
+            return -1
+
+        return 1 - spatial.distance.cosine(v1, v2)
+
+    def find_golden(self, query, wm: list) -> set:
+
+        ## Calculate score of query for each para
+        scores = [
+            (ith, self.get_score(wm[query], wm[ith]))
+            for ith in range(0, len(wm) - 3)
+        ]
+
+        goldens   = set()
+
+        for score in scores:
+            if score[1] > 0:
+                goldens.add(score[0])
+                if len(goldens) == args.n_paras:
+                    break
+
+        if len(goldens) < args.n_paras:
+            for score in scores:
+                if score[1] == 0:
+                    goldens.add(score[0])
+                    if len(goldens) == args.n_paras:
+                        break
+
+        return goldens
+
+    def process_entry(self, entry):
+        """This function is used to run in parallel to read and initially preprocess
+        raw documents. Afterward, it puts a dict containing result into queue.
+
+        Args:
+            document (dict): a document of dataset
+        """
+        #########################
+        ## Preprocess context
+        #########################
+        paras   = self.processed_contx[entry['document']['id']]
+
+
+        #########################
+        ## Preprocess question and answer
+        #########################
+        ques    = self.process_ques_ans(entry['question']['text'])
+        ans1    = self.process_ques_ans(entry['answers'][0]['text'])
+        answ2   = self.process_ques_ans(entry['answers'][1]['text'])
+
+
+        #########################
+        ## TfIdf vectorize
+        #########################
+        tfidfvectorizer = TfidfVectorizer(analyzer='word', stop_words='english',
+                                            ngram_range=(1, 3), max_features=500000)
+
+        wm = tfidfvectorizer.fit_transform(paras + [ques, ans1, answ2]).toarray()
 
         question, answer1, answer2 = len(wm) - 3, len(wm) - 2, len(wm) - 1
 
-        neighbor_paras_ques = {question}
-        find_golden(question, wm, 0, neighbor_paras_ques)
-        neighbor_paras_ques.discard(question)
 
-        neighbor_paras_answ1 = {answer1}
-        find_golden(answer1, wm, 0, neighbor_paras_answ1)
-        neighbor_paras_answ2 = {answer2}
-        find_golden(answer2, wm, 0, neighbor_paras_answ2)
+        #########################
+        ## Find golden paragraphs from
+        ## question and answers
+        #########################
+        golden_paras_ques   = self.find_golden(question, wm)
 
-        neighbor_paras_answ = neighbor_paras_answ1 | neighbor_paras_answ2
-        neighbor_paras_answ.discard(answer1)
-        neighbor_paras_answ.discard(answer2)
+        golden_paras_answ1  = self.find_golden(answer1, wm)
+        golden_paras_answ2  = self.find_golden(answer2, wm)
+        golden_paras_answ   = golden_paras_answ1 | golden_paras_answ2
 
 
-        queue.put({
-            'doc_id'    : document['document']['id'],
-            'question'  : ' '.join(document['question']['tokens']),
-            'answers'   : [' '.join(x['tokens']) for x in document['answers']],
-            'En'        : [paras[i] for i in neighbor_paras_answ],
-            'Hn'        : [paras[i] for i in neighbor_paras_ques]
-        })
+
+        return {
+            'doc_id'    : entry['document']['id'],
+            'question'  : ' '.join(entry['question']['tokens']),
+            'answers'   : [' '.join(x['tokens']) for x in entry['answers']],
+            'En'        : [paras[i] for i in golden_paras_answ],
+            'Hn'        : [paras[i] for i in golden_paras_ques]
+        }
 
 
-def trigger_reading_data():
-    """ Start reading and processing data
-    """
+    def trigger_reading_data(self):
+        """ Start reading and processing data
+        """
+        #########################
+        # Process contexts first
+        #########################
+        for entry in load_dataset('narrativeqa', split=self.split):
+            try:
+                id_ = entry['document']['id']
+                _   = self.processed_contx[id_]
+            except KeyError:
+                self.processed_contx[id_] = np.array([entry['document']['text'],
+                                                      entry['document']['kind'],
+                                                      entry['document']['start'],
+                                                      entry['document']['end']])
 
-    for split in ['validation', 'train', 'test']:
+        ParallelHelper(self.f_process_text, self.processed_contx.keys(),
+                       lambda d, lo, hi: d[lo, hi],
+                       num_proc=args.num_proc).launch()
+
+
+        #########################
+        # Process each entry of dataset
+        #########################
         for shard in range(8):
             ### Need to check whether this shard has already been processed
-            path    = PATH['dataset_para'].replace("[SPLIT]", split).replace("[SHARD]", str(shard))
+            path    = PATH['dataset_para'].replace("[SPLIT]", self.split).replace("[SHARD]", str(shard))
             if check_file_existence(path):
                 continue
 
 
-            logging.info(f"= Preprocess dataset: {split} - shard {shard}")
+            logging.info(f"= Process dataset: {self.split} - shard {shard}")
 
-            dataset = load_dataset('narrativeqa', split=split).shard(8, shard)
+            dataset = load_dataset('narrativeqa', split=self.split).shard(8, shard)
 
-            list_documents  = ParallelHelper(f_trigger, dataset,
-                                             lambda d, lo, hi: d.select(range(lo, hi)),
-                                             num_proc=args.num_proc).launch()
+            list_documents = [self.process_entry(entry) for entry in dataset]
 
             save_object(path, pd.DataFrame(list_documents))
 
@@ -192,4 +353,5 @@ def trigger_reading_data():
 if __name__ == '__main__':
     logging.info("* Reading raw data and decompose into paragraphs")
 
-    trigger_reading_data()
+    for split in ['validation', 'train', 'test']:
+        DataReading(split).trigger_reading_data()